Fixes VSTS Bug 1027415: [FATAL] SigTerm signal in MonoDevelop.VersionControl.Git.dll!MonoDevelop.VersionControl.Git.GitRepository::Dispose+49
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1027415

Hard to reproduce the issue (only 4 Hits). The GitConfigurationDialog
does some long running operations where the cancellation token isn't
passed to these. That got fixed.

But the main issue is the Dispose that could dead lock on the
GitRepository. Since destroy is done on the dedicated operation thread
it may dead lock in some situations.
That can only happen if something long running blocks the
DedicatedOperationFactory. Now the DedicatedOperationFactory
operations get linked to the disposeToken which should cancel these
operations.
But all actions need to get the combined token and use that for their
cancellations.